requirements.txt
aiofiles==25.1.0
anyio==4.11.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
fastapi==0.110.1
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
idna==3.11
joblib==1.5.2
nltk==3.8.1
packaging==25.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==1.10.24
pyparsing==3.2.5
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.6
pytokens==0.3.0
regex==2025.11.3
requests==2.31.0
rsa==4.9.1
sniffio==1.3.1
starlette==0.37.2
tqdm==4.67.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.29.0